Began refactoring to split boundary timer event tests from common ones.
